setBattleship Flowchart (#9)
This is a draft of how to set up the battleships in the game. 
--------------------------------------------------------------------
Variables
int size:the length of the battleship
string orientation: sets whether the ship is horizontal or vertical
Board board[][]: the board for the game
int xCoordinate: the starting x position of the battleship
int yCoordinate: the starting y position of the battleship
-------------------------------------------------------------------------
I did not include a check method within this function.
It could probably be called as a separate function to 
ensure that there are no overlapping battleships.
User input would also need verification so that the function
will not exit prematurely.